+++ route_builder.go
@@ -7,6 +7,7 @@ package restful
 import (
 	"log"
 	"reflect"
+	"runtime"
 	"strings"
 )
 
@@ -126,6 +127,7 @@ func (b *RouteBuilder) Param(parameter *Parameter) *RouteBuilder {
 }
 
 // Operation allows you to document what the acutal method/function call is of the Route.
+// Unless called, the operation name is derived from the RouteFunction set using To(..).
 func (b *RouteBuilder) Operation(name string) *RouteBuilder {
 	b.operation = name
 	return b
@@ -191,6 +193,11 @@ func (b *RouteBuilder) Build() Route {
 	if b.function == nil {
 		log.Fatalf("[restful] No function specified for route:" + b.currentPath)
 	}
+	operationName := b.operation
+	if len(operationName) == 0 && b.function != nil {
+		// extract from definition
+		operationName = nameOfFunction(b.function)
+	}
 	route := Route{
 		Method:         b.httpMethod,
 		Path:           concatPath(b.rootPath, b.currentPath),
@@ -202,7 +209,7 @@ func (b *RouteBuilder) Build() Route {
 		pathExpr:       pathExpr,
 		Doc:            b.doc,
 		Notes:          b.notes,
-		Operation:      b.operation,
+		Operation:      operationName,
 		ParameterDocs:  b.parameters,
 		ResponseErrors: b.errorMap,
 		ReadSample:     b.readSample,
@@ -213,4 +220,17 @@ func (b *RouteBuilder) Build() Route {
 
 func concatPath(path1, path2 string) string {
 	return strings.TrimRight(path1, "/") + "/" + strings.TrimLeft(path2, "/")
+}
+
+// nameOfFunction returns the short name of the function f for documentation.
+// It uses a runtime feature for debugging ; its value may change for later Go versions.
+func nameOfFunction(f interface{}) string {
+	fun := runtime.FuncForPC(reflect.ValueOf(f).Pointer())
+	tokenized := strings.Split(fun.Name(), ".")
+	last := tokenized[len(tokenized)-1]
+	last = strings.TrimSuffix(last, ")·fm") // < Go 1.5
+	last = strings.TrimSuffix(last, ")-fm") // Go 1.5
+	last = strings.TrimSuffix(last, "·fm")  // < Go 1.5
+	last = strings.TrimSuffix(last, "-fm")  // Go 1.5
+	return last
 }